ci: split lotus-ai into explicit CI lanes
