+++ .github/workflows/pr-merge-gate.yml
@@ -1,9 +1,8 @@
-name: Backend Service Pipeline v2
+name: Pull Request Merge Gate
 
 on:
-  push:
   pull_request:
-    branches: ["main"]
+    branches: [ "main" ]
   workflow_dispatch:
 
 concurrency:
@@ -17,17 +16,18 @@ env:
   PIP_DISABLE_PIP_VERSION_CHECK: "1"
   PYTHONUNBUFFERED: "1"
   PYTHON_VERSION: "3.12"
+  COVERAGE_FAIL_UNDER: "99"
 
 jobs:
   workflow-lint:
-    name: Workflow Lint
+    name: PR Merge Gate / Workflow Lint
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
       - uses: reviewdog/action-actionlint@v1
 
   lint-typecheck-security:
-    name: Lint Typecheck Security
+    name: PR Merge Gate / Lint Typecheck Security
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
@@ -38,7 +38,7 @@ jobs:
       - name: Install
         run: make install
       - name: Verify Dependencies
-        run: python -m pip check
+        run: make verify-dependencies
       - name: Lint
         run: make lint
       - name: Typecheck
@@ -59,7 +59,7 @@ jobs:
         run: make security-audit
 
   test-suites:
-    name: Tests (${{ matrix.suite }})
+    name: PR Merge Gate / Tests (${{ matrix.suite }})
     needs: [lint-typecheck-security]
     runs-on: ubuntu-latest
     strategy:
@@ -98,7 +98,7 @@ jobs:
           if-no-files-found: error
 
   runtime-mode-smoke:
-    name: Runtime Mode Smoke
+    name: PR Merge Gate / Runtime Mode Smoke
     needs: [lint-typecheck-security]
     runs-on: ubuntu-latest
     steps:
@@ -113,7 +113,7 @@ jobs:
         run: make runtime-mode-smoke
 
   coverage-gate:
-    name: Coverage Gate (Combined)
+    name: PR Merge Gate / Coverage Gate (Combined)
     needs: [test-suites, runtime-mode-smoke]
     runs-on: ubuntu-latest
     steps:
@@ -130,13 +130,13 @@ jobs:
           pattern: coverage-data-*
           merge-multiple: true
           path: coverage-data
-      - name: Enforce 99% coverage
+      - name: Enforce coverage floor
         run: |
           python -m coverage combine coverage-data
-          python -m coverage report --fail-under=99
+          python -m coverage report --fail-under=${{ env.COVERAGE_FAIL_UNDER }}
 
   docker-build:
-    name: Validate Docker Build
+    name: PR Merge Gate / Validate Docker Build
     needs: [coverage-gate]
     runs-on: ubuntu-latest
     steps:
